requirements.txt
# Bronze Tier
watchdog>=4.0.0

# Silver Tier — Gmail Watcher
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Silver Tier — LinkedIn Watcher
playwright>=1.40.0
requests>=2.31.0

# Gold Tier — Twitter/X Watcher
tweepy>=4.14.0

# Gold Tier — Facebook/Instagram Watcher
# (uses requests, already included)

# Gold Tier — Odoo Integration
# (uses urllib, built-in)

# Gold Tier — Social Media MCP
# node-fetch (npm package, installed via npm)

# Shared
python-dotenv>=1.0.0

# Gold Tier — Optional enhancements
# retry>=0.9.0
# circuitbreaker>=2.0.0
